backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "frontier-ai-radar"
version = "0.1.0"
description = "Daily Multi-Agent Intelligence System - Frontier AI Radar"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "feedparser>=6.0.0",
    "playwright>=1.41.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "prefect>=2.14.0",
    "openai>=1.12.0",
    "reportlab>=4.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "apscheduler>=3.10.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
